% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_setup.R
\name{run_setup}
\alias{run_setup}
\title{Run Setup}
\usage{
run_setup(
  conn_fun = "pg13::local_connect()",
  civic_schema = "civic",
  log_table_name = "setup_civic_log",
  civic_version = as.character(Sys.time()),
  verbose = TRUE,
  render_sql = TRUE
)
}
\arguments{
\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect()'}

\item{civic_schema}{PARAM_DESCRIPTION, Default: 'civic'}

\item{log_table_name}{PARAM_DESCRIPTION, Default: 'setup_civic_log'}

\item{civic_version}{PARAM_DESCRIPTION, Default: as.character(Sys.time())}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{render_sql}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
Run Setup
}
